casted rate to a double

+++ lib/pages/search_therapists_page.dart
@@ -73,7 +73,7 @@ class _SearchTherapistsPageState extends State<SearchTherapistsPage> {
                               name: document['name'],
                               imagePath: document['imagePath'],
                               path: document['path'],
-                              rate: document['rate'],
+                              rate: document['rate'].toDouble(),
                               rating: document['rating'].toDouble(),
                               type: document['type']);
